receitas/python/TSE_Resultados/requirements.txt
selenium>=3.141.0
Appium-Python-Client>=1.0.2
pytest>=6.1.2